Move tick formatter to a separate file
